modifed kctl to support environments where we must use nat's and port forwards to connect the external network to the site network (eg. challenger and intrepid at anl)
in support of above added a script that creates port forwards for ssh to a particular compute node via its io node (nat server)

added to the repo khbootapp that was developed to automatically construct
various cluster like confgurations.  Has been out of a repo for a while
finally getting it in.